pyproject.toml
[tool.poetry]
name = "flower_shop"
version = "0.1.0"
description = ""
authors = ["Nikita Baranov <nbaranov.vrn@gmail.com> Alexander Korzh <alex@korzh.me>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.8"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
